Update to latest MSBuild.Community.Tasks
Remove a bunch of Windows specific build code now that MS build community tasks runs on Mono 4

Remove requirement that versions start with "v"

Allow publishing from non-master branches (beta releases, for instance)

revamp conditions

Fix Mono build

fix push

fix nuget push

version string

2.5.0 Release Notes

No need to skip tests on Mono now. Modify configuration to use riak-client-tools default port of 8087

Remove SkipMono arg from nunit runner

Add test for ListKey with Blob Key Cell & fetching that row

Move restoring NuGet packages into the common build.targets file.

Update NUnit to 3.5.0

Use TsColumnDescription to set ColumnType on a Cell

Fix paths to certificate files.

update tools to 1.3.4

Fix Mono build

Replace -exclude with --where

Add integration Hll target

+++ src/Test/Unit/CRDT/UpdateSetTests.cs
@@ -75,9 +75,9 @@ namespace Test.Unit.CRDT
             Assert.AreEqual(Encoding.UTF8.GetBytes(BucketType), protobuf.type);
             Assert.AreEqual(Encoding.UTF8.GetBytes(Bucket), protobuf.bucket);
             Assert.AreEqual(Encoding.UTF8.GetBytes(Key), protobuf.key);
-            Assert.AreEqual(q3, protobuf.w);
-            Assert.AreEqual(q1, protobuf.pw);
-            Assert.AreEqual(q2, protobuf.dw);
+            Assert.AreEqual((uint)q3, protobuf.w);
+            Assert.AreEqual((uint)q1, protobuf.pw);
+            Assert.AreEqual((uint)q2, protobuf.dw);
             Assert.IsTrue(protobuf.return_body);
             Assert.IsFalse(protobuf.include_context);
             Assert.AreEqual(20000, protobuf.timeout);
